3.6 Level Grid
This lecture highlights the problem of not defining the game's requirements before designing the structure of the code. The instructor failed to mention how the system would be implemented, and it is leading to a very passive learning style with no room for exploration (if I change something, it may be incorrect and will cause massive headaches later down the run).

I will attempt to learn new concepts within this course, but the game is as good as dead without properly highlighting its requirements. It's just too much confusion.