5.2 Enemy
Basic setup for getting the enemy unit system logic ready. An enemy AI script was created, and it currently bounces back the turn to the player. A prefab variant of the unit is generated for the enemy unit, and the enemy unit shares the same unit script as the player. the only difference is the enemy unit has a bool marked true for isEnemy.

Other changes in this commit include updating the UI visual to handle enemy turn logic and proper action points resetting for players and enemies.