5.3 Shoot Action
This commit handles adding a shoot action to the units. Most of the code is copied from other actions (which begs the question if some generic methods should be created). One small difference from this commit and the courses is that the unit's position is being exposed in the lecture, but we already have the logic for actions to receive the unit's position already, so exposing the position is redundant and was not done in this commit.