5.4 Unit Animator
This commit introduces the unit animator class. This class will handle the animation side of actions and other events. However, the implementation in this lecture is flawed. Not every action has an animation, but every animation dependent on another script has a corresponding action.

The dependency relationship from action to the animator is  0,M<->1,1; each action references one animator, and each animator references 0 or more actions. Thus, the dependency should be on the actions, not the animator, that way we don't need to handle TryGetComponent methods (which can be confusing if the game object doesn't have a specific action), and all the logic for animations can be easily read in the respective scripts.

By the end of this course, this code will be heavily refactored. For the purposes of this course, I will follow along with what the instructor does.