5.12 Enemy AI Complex
Key Takeaway:

In this game, the AI behavior is calculated by a score metric. The better the score, the more likely that the enemy AI will use the action related to that score. All possible actions are computed for the enemy AI, meaning every movement position, shoot action, and spin action, are computed (spin actions have 0% change of ever happening). This may be slow if the game were to be expanded upon, but for now it works well.

There is room to expand on enemy movement logic. For example, right now, enemies will move to be in range of a target so they can shoot them. However, they are not yet smart enough to understand unit range, meaning they move to a position that is less reasonable if their goal is to hit one unit and be in range of as little units as possible. A score value for each grid position can be computed in the future to consider the damage they may take while at a given position.