6.6 Pathfinding Implementation
There are small differences in the effect of the algorithm I have and the one tested in the course. The code should be the same, but in testing the results are slightly different. Regardless, the current system does find the most optimal path to get to a node.

This pathfinding implements the A* algorithm, but leaves out a weight-based movement (all nodes are equally weighed the same). In future sections, the system will be expanded upon to prevent unwalkable path nodes.