7.9 Interface IInteractable
This section is confusing. The purpose of an interface is described to implement blueprints that a class must follow, but the instructor's implementation leaves duplicate code between IInteractible classes that feels like it should be more of an inherited class instead of an interface.

An online search suggests that an interface is the proper methodology here, since interactions will vary vastly based on what the interaction is meant to do. This course just used two very similar setups that made it confusing.