7.11 Input System
The input system is now refactored to work both with the legacy input system and the new input system.

To switch between a new and old input system, the inputmanager script declares a variable using #declare. This can be used to test if the script declares such a variable or not, switching the script's logic accordingly. This is commonly used by Unity.

For the new input system, some of the input is grabbed from the pre-built input system, but the rest is generated from a basic C# class auto-generated by the player input actions map. The InputManager creates an instance of this class and uses that to read the player input.