Implement MonobehaviourEventListener abstract class
The MonobehaviourEventListener class sets up a simple structure for classes that want to inherit from MonoBehaviour and have subscribed events to force cleanup of subscriptions on disable and destroy.

Future commits will focus on refactoring code to follow better practices, such as reducing variable declaration in Update methods and serializing variables that can and should be custom tailored.